+++ app/controllers/challenges_controller.rb
@@ -1,17 +1,20 @@
 # frozen_string_literal: true
 
+# Controller for challenge listings detail page and contact form handling
 class ChallengesController < ApplicationController
+  include ChallengesHelper
+
   def show
-    @challenge = Challenge.includes(phases: {phase_winner: :winners})
-                          .find_by(custom_url: params[:challenge])
+    @challenge = Challenge.includes(phases: { phase_winner: :winners }).
+      find_by(custom_url: params[:challenge])
     @section = params[:section] || 'overview'
 
     if @challenge.nil?
-      redirect_to '/', alert: "Challenge not found."
+      redirect_to '/', alert: t('challenge_listing.alerts.error')
       return
     end
 
-    @logo_url = determine_logo_url
+    @logo_url = challenge_logo_url(@challenge)
     render :show
   end
 
@@ -21,7 +24,7 @@ class ChallengesController < ApplicationController
     result = ContactFormsService.send_email(@challenge, contact_form_params)
 
     if result[:success]
-      flash[:notice] = "Your message has been sent successfully."
+      flash[:notice] = t('mailers.contact_form.sent_successfully')
     else
       flash[:error] = result[:errors].join(", ")
     end
@@ -31,32 +34,12 @@ class ChallengesController < ApplicationController
 
   private
 
-  def determine_logo_url
-    return nil unless @challenge.upload_logo
-    return nil unless @challenge.logo_key.present?
-
-    if Rails.env.production? || Rails.env.staging?
-      s3_logo_url(@challenge)
-    else
-      file_system_logo_url(@challenge)
-    end
-  end
-
-  def file_system_logo_url(challenge)
-    "#{ENV.fetch('PHOENIX_URI')}/uploads/challenges/original-#{challenge.logo_key}#{challenge.logo_extension}"
-  end
-
-  def s3_logo_url(challenge)
-    "#{ENV.fetch('S3_BASE_URL')}/challenges/original-#{challenge.logo_key}#{challenge.logo_extension}"
-  end
-
   # contact form
   def find_challenge
-    @challenge = Challenge.find_by(custom_url: params[:challenge]) ||
-                 Challenge.find(params[:challenge])
+    @challenge = Challenge.find_by(custom_url: params[:challenge]) || Challenge.find(params[:challenge])
 
     if @challenge.nil?
-      flash[:error] = "Challenge not found"
+      flash[:error] = t('challenge_listing.alerts.error')
       redirect_to challenges_path
       return false
     end
