+++ app/views/challenges/show.html.erb
@@ -1,4 +1,4 @@
-<div class="challenge-show-page" data-controller="challenge-detail modal"">
+<div class="margin-top-4 challenge-show-page" data-controller="challenge-detail modal"">
   <div class="radius-lg usa-card usa-card__container">
     <div class="padding-top-4 usa-card__header">
       <div class="usa-grid">
@@ -115,7 +115,7 @@
                   <p class="font-sans-md margin-0 mobile-lg:margin-1"><%= number_to_currency(@challenge.prize_total / 100) %></p>
                 </div>
               </div>
-              <div class="grid-col-12">
+              <div class="grid-col-4">
                 <%= link_to "#", class: "usa-link display-flex flex-align-center text-no-wrap", onclick: "window.print(); return false;" do %>
                   <%= image_tag('images/usa-icons/print.svg', class: "usa-icon usa-icon--size-3 margin-right-1 icon-primary-orange", alt: "print") %>
                   <span class="font-sans-md">Print challenge</span>
